refactor(architecture): move completion components to feature-based structure (#46)
- Move HabitCompletionCard and DailyHabitList from shared/ to features/completion/
- Follow feature-based architecture guidelines from frontend/CLAUDE.md
- Remove premature abstraction - components were only used by completion feature
- Update import paths and maintain co-located tests
- Clean up shared/components exports to only include truly shared components
- Preserve all functionality while improving maintainability and code organization
