+++ frontend/src/features/completion/components/__tests__/DailyHabitList.test.tsx
@@ -2,7 +2,7 @@ import { describe, it, expect, vi } from 'vitest'
 import { render, screen } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import { DailyHabitList } from '../DailyHabitList'
-import type { Habit } from '../../../types/data'
+import type { Habit } from '../../../../types/data'
 
 const mockHabits: Habit[] = [
   {
