+++ frontend/src/shared/components/index.ts
@@ -9,9 +9,7 @@ import { lazy } from 'react';
 export const Modal = lazy(() => import('./Modal').then(m => ({ default: m.Modal })));
 export const ConfirmModal = lazy(() => import('./ConfirmModal').then(m => ({ default: m.ConfirmModal })));
 
-// Daily completion components (not lazy - core to shutdown ritual)
-export { HabitCompletionCard } from './HabitCompletionCard';
-export { DailyHabitList } from './DailyHabitList';
+// Shared UI components
 export { ProgressIndicator } from './ProgressIndicator';
 
 // Export types for TypeScript
